feat: refactor settings management to use new settings panel composables and remove deprecated editor settings
