+++ src/renderer/features/settings/composables/useSettingsAppearance.ts
@@ -1,9 +1,8 @@
-import { watch, onMounted, onUnmounted } from 'vue';
+import { onMounted, onUnmounted, watch } from 'vue';
 import { useSettingsStore, type AccentMode, type ThemeMode } from '../store/settings.store';
 
-export function useGeneralSettings() {
+export function useSettingsAppearance() {
   const settingsStore = useSettingsStore();
-  let mediaQuery: MediaQueryList | null = null;
   let removeSystemThemeListener: (() => void) | null = null;
 
   const resolveThemeMode = (mode: ThemeMode): 'light' | 'dark' => {
@@ -29,10 +28,21 @@ export function useGeneralSettings() {
     }
   };
 
+  const applyEditorTypography = (fontSize: number, fontFamily: string) => {
+    const root = document.documentElement;
+    root.style.setProperty('--editor-font-size', `${fontSize}px`);
+
+    if (fontFamily) {
+      root.style.setProperty('--editor-font-family', fontFamily);
+    } else {
+      root.style.removeProperty('--editor-font-family');
+    }
+  };
+
   watch(
     () => [settingsStore.config.themeMode, settingsStore.config.accentMode] as const,
-    ([newThemeMode, newAccentMode]) => {
-      applyThemeAppearance(newThemeMode, newAccentMode);
+    ([themeMode, accentMode]) => {
+      applyThemeAppearance(themeMode, accentMode);
     },
     { immediate: true },
   );
@@ -45,8 +55,16 @@ export function useGeneralSettings() {
     { immediate: true },
   );
 
+  watch(
+    () => [settingsStore.config.editorFontSize, settingsStore.config.editorFont] as const,
+    ([fontSize, fontFamily]) => {
+      applyEditorTypography(fontSize, fontFamily);
+    },
+    { immediate: true },
+  );
+
   onMounted(() => {
-    mediaQuery = window.matchMedia('(prefers-color-scheme: dark)');
+    const mediaQuery = window.matchMedia('(prefers-color-scheme: dark)');
     const handleSystemThemeChange = () => {
       if (settingsStore.config.themeMode === 'system') {
         applyThemeAppearance('system', settingsStore.config.accentMode);
@@ -55,7 +73,7 @@ export function useGeneralSettings() {
 
     mediaQuery.addEventListener('change', handleSystemThemeChange);
     removeSystemThemeListener = () => {
-      mediaQuery?.removeEventListener('change', handleSystemThemeChange);
+      mediaQuery.removeEventListener('change', handleSystemThemeChange);
     };
   });
